requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1
snowflake-connector-python==3.12.4
pandas==2.2.3
snowflake-connector-python==3.12.4
boto3==1.36.2